Fix: various v3 improvements

+++ Package/Runtime/CrashKonijn.Goap.Core/Interfaces/IActionConfig.cs
@@ -4,7 +4,7 @@ namespace CrashKonijn.Goap.Core
 {
     public interface IActionConfig : IClassConfig
     {
-        int BaseCost { get; }
+        float BaseCost { get; }
         ITargetKey Target { get; }
         float StoppingDistance { get; }
         bool RequiresTarget { get; }
